Address Aaron's comments
